requirements-render.txt
flask==2.3.3
flask-cors==4.0.0
chromadb==0.4.15
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-docx==0.8.11
gunicorn==21.2.0
numpy==1.24.3
scipy==1.10.1
scikit-learn==1.3.0
pandas==2.0.3
openpyxl==3.1.2
